Fix bug in how noise scale is computed
Per Equation 15 in [the paper](https://arxiv.org/pdf/2102.11673.pdf), the relation between the noise parameter sigma (`scale`) and the spectral norm of the Jacobian (`eta_max`) should involve division by `eta`, not multiplication. Simply put: to guarantee a lower FIL, you need to add noise with a larger scale.

This corrects that mistake in the code.